+++ src/Client/ManagesReadWriteTrait.php
@@ -42,6 +42,12 @@ trait ManagesReadWriteTrait
      */
     public function loadGeneratedTypes(GeneratedTypeRegistrar $registrar): self
     {
+        if (! property_exists($registrar, 'only') || ! $registrar->only) {
+            foreach ($registrar->dependencyRegistrars() as $dependency) {
+                $this->loadGeneratedTypes($dependency);
+            }
+        }
+
         $registrar->registerCodecs($this->repository);
         $this->enumMappings = array_merge($this->enumMappings, $registrar->getEnumMappings());